Refactor of template class hierarchy
The `Skeleton` classes are gone: they were an unnecessary complication
and their function is now contained in the `Template` types.

There is a new field `TemplateConfig` in the `ResumeSchema`'s `Meta`
subfield. This is a collection of *string dict keys* of configuration
options for the template.

We cannot directly use e.g. `Template`, `Theme` etc. in `resume.py`
since `Template` itself relies on `ResumeSchema`: a circular import
results. As a solution, let `TemplateConfig` accept strings and use
these as keys into a dictionary of all available templates, themes and
translations later on.
This way, `ResumeSchema` doesn't need to know anything about any
`Template` and such, greatly simplifying the code.

Using the string keys, we can use `from_model_config` to create a proper
`Template` object.

Through the `render` method, a fully instantiated template now knows
how to render itself.

Themes are now *instances* instead of concrete classes inheriting from
an abstract base class. This is a good reduction in complexity and
plain lines of code as well.