feat(build): Halve the Docker image size
Achieved by downgrading the base image and instead installing some
dependencies manually.

This is still awful: hundreds of MB of build tools are contained in the final image.
However, building the Python application/dependencies, then copying stuff around in a multi-stage
build is a nightmare and not worth the effort for now.