feat(app): Add special `heyho` endpoint for showcasing
This is excellent: it allows anyone to try out ancv immediately,
without me having to provide my own resume (which seems selfish).
This is a neutral sample. It can also serve as a starting point for constructing
a resume.

Let's hope the 'heyho' GitHub user doesn't fancy using ancv!
(that account is taken but completely empty as of right now)