feat: Add command to generate JSON schema for the package
There's a fundamental issue: the app works off of pydantic models. They
are the source of truth. We'd like to provide an ancv-specific JSON
Schema for users to fill out their JSONs against/with. IDEs like VSCode
would then help with auto-completion. Nice win!

Ideally, we'd like to auto-generate the JSON Schema from the pydantic
models. This has the downside of suddenly including a verbose version of
the (licensed) JSON Resume schema instead of just referencing it via
URL. It also leads to inconsistencies.

We could also hand-write a JSON Schema and auto-generate the Pydantic
models from it (the inverse way). This has countless more problems, like
the weird ISO-8601 regex-based definition in the JSON Resume Schema,
which we had changed to `date` types in the pydantic models.

As such, it was deemed most suitable to hand-write *both* the JSON
Schema for ancv as well as the pydantic models. This leads to (some
minor) code duplication and the risk of the two becoming asynchronous.
If this proves a pain in the future, look into this issue again. For
now, the pain of keeping the two in sync is much greater.