ci: Remove `locked`, fall back to `frozen`
This JSON Path manipulation crap does not work
with release-please, although it works locally,
manually.

The problem is that `uv.lock` will contain our
package version number and will NOT be updated in
release please PRs, leading to an out of date
`uv.lock`, which needs to be updated manually with
`uv lock` before `--locked` will start working.
Else, running with `--locked` fails.

We cannot easily run `uv lock` in CI, so fall back
to `--frozen`.